rename and merge Encoding.h to encoding.h
